zenith-backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Vector Database - Pinecone
pinecone-client==3.0.0
pyreadline3  # Windows readline support

# LLM Providers - Using Local LM Studio (OpenAI-compatible API)
# No cloud API keys needed! 🎉
# openai==1.10.0  # NOT NEEDED - Using local LM Studio
# google-generativeai==0.3.2  # NOT NEEDED - Using local LM Studio

# PDF Generation
reportlab==4.0.9
pypdf2==3.0.1

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Text Processing - Using Local LLM (removed cloud dependencies)
# langchain==0.1.4  # NOT NEEDED - Direct API calls to LM Studio
# langchain-openai==0.0.5  # NOT NEEDED
# langchain-community==0.0.16  # NOT NEEDED  
# tiktoken==0.5.2  # NOT NEEDED - LM Studio handles tokenization

# Monitoring & Logging
python-json-logger==2.0.7
